Type Alias span::Span

pub type Span = SpanData<SyntaxContextId>;

Aliased Type

struct Span {
    pub range: TextRange,
    pub anchor: SpanAnchor,
    pub ctx: SyntaxContextId,
}

Fields
range: TextRange

The text range of this span, relative to the anchor. We need the anchor for incrementality, as storing absolute ranges will require recomputation on every change in a file at all times.

anchor: SpanAnchor

The anchor this span is relative to.

ctx: SyntaxContextId

The syntax context of the span.

Implementations

impl Span

pub const DUMMY: Self = _

👎Deprecated: dummy spans will panic if surfaced incorrectly, as such they should be replaced appropriately