Crate span

File and span related types.

Structs

FileId

Handle to a file in [Vfs]

FilePosition

FileRange

HirFileId

Input to the analyzer is a set of files, where each file is identified by FileId and contains source code. However, another source of source code in Rust are macros: each macro can be thought of as producing a “temporary file”. To assign an id to such a file, we use the id of the macro call that produced the file. So, a HirFileId is either a FileId (source code written by user), or a MacroCallId (source code produced by macro).

MacroCallId

MacroCallId identifies a particular macro invocation, like println!("Hello, {}", world).

MacroFileId

RealSpanMap

SpanAnchor

SpanData

SpanMap

Maps absolute text ranges for the corresponding file to the relevant span data.

SyntaxContextId

TextRange

A range in text, represented as a pair of TextSize.

TextSize

A measure of text length. Also, equivalently, an index into text.

Enums

HirFileIdRepr

Constants

FIXUP_ERASED_FILE_AST_ID_MARKER

FileId used as the span for syntax node fixups. Any Span containing this file id is to be considered fake.

ROOT_ERASED_FILE_AST_ID

Type Aliases

ErasedFileAstId

Span